requirements.txt
tweepy==4.14.0
python-dotenv==0.19.0
logging==0.4.9.6